Align cdada_xxx_create()/create_custom()
Change `cdada_xxx_create()` to a MACRO that receives a TYPE to be
consistent with `cdada_xxx_create_custom()`, which it _has_ to be
a MACRO.

Now `cdada_xxx_create( TYPE )` receives a type, and the `sizeof()`
is automatically done by the MACRO expansion.

Also fix some non up-to-date documentation of these methods.