benchmark: Condtionalize AMD64 Time Stamp Counter
Use a Time Stamp Counter only on AMD64 based platforms. Other platforms
will have the ability to run the benchmark, but zeros will be reported
as time values. A warning is printed at compile time to notify the user
that Time Stamp Counters are not available for non-AMD64 based platforms.