test/list|set|map: properly test destructor
Make sure destructor is tested with elements in container to
detect leaks. This was not important when container's memory (the
container itself) was dynamically allocated, but now that it isn't,
the tests should ensure containers have at least one element so
that memory leaks (due to not calling destructor) are properly
detected by the test harness.