map: set ptr_not_null to a valid pointer
ptr_not_null should contain a valid pointer, otherwise the map_test can
segfault. Common patter used in __map_internal.h is

  T* aux;
  aux = (T*)key;
  it = m_u->find(*aux);

The key is set to (void*)0x123 in multiple test_basics test cases. It is
cast to (int*) and dereferenced in find method call. This is the place
where map_test can segfault as it cannot access memory at address 0x123.

This issue is fixed by using a real accessible pointer to an integer
with value 0x123.